% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_donut_fe.R
\name{basic_donut_fe}
\alias{basic_donut_fe}
\title{Perform a Donut FE analysis}
\usage{
basic_donut_fe(dist, ds, dep_var, indep_vars, fe = "id")
}
\arguments{
\item{dist}{A vector with two distances: the inner and the outer.
The inner distance refers to the maximum distance for treatment.
The outer distance refers to the overall population: it is the cut-off value.}

\item{ds}{The dataset}

\item{dep_var}{The dependent variable in the regression}

\item{indep_vars}{a character vector of independent variables. Must be part of \code{ds}.}

\item{fe}{a character vector of Fixed Effects variables. Must be part of \code{ds}.}
}
\value{
A single plm model.
}
\description{
Performs a single FE-models with a given data set.
The independent variable and the population vary.
}
\keyword{internal}
