% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_df_to_qmd_table.R
\name{convert_to_qmd_table}
\alias{convert_to_qmd_table}
\title{Data frame to qmd table}
\usage{
convert_to_qmd_table(df, column_names = names(df)[-ncol(df)])
}
\arguments{
\item{df}{The data frame that should be converted}

\item{column_names}{The column names that should be printed.
Defaults to \code{names(df)}.}
}
\value{
A qmd table.
}
\description{
Converts a data frame or tibble to a qmd table.
This is handy for \code{qmd} Output.
Note: if you want to use this, you need to specify the following:
\code{#| eval: asis} in the R chunk.
}
\examples{
convert_to_qmd_table(iris[1:10,])
}
