% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_fe.R
\name{donut_analysis}
\alias{donut_analysis}
\title{Perform a Donut FE analysis}
\usage{
donut_analysis()
}
\arguments{
\item{dist}{A vector with two distances: the inner and the outer.}
}
\value{
A a list of plm models.
}
\description{
Performs multiple FE-models with a given data set.
The independent variable and the population vary.
}
