% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_models.R
\name{donut_models}
\alias{donut_models}
\title{Perform a Donut FE analysis}
\usage{
donut_models(inner = 3:9, outer = (1:5) * 10, ds, indep_vars, ...)
}
\arguments{
\item{inner}{A vector with the inner radii.}

\item{outer}{A vector with the outer radii}

\item{ds}{The dataset}

\item{indep_vars}{The independent variables to pass on to \code{donut_analysis}}

\item{...}{Additional arguments}
}
\value{
A donut_models list: \code{donut_list}
}
\description{
Performs a single FE-models with a given data set.
The independent variable and the population can be adjusted.
}
