% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donutsR.R
\docType{package}
\name{donutsR}
\alias{donutsR}
\alias{donutsR-package}
\title{Running donut analyses}
\description{
With this package, you can run donut analyses: regressions with a distance parameter, which is used to create treatment dummies, and to determine the population.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sarahzeller.github.io/donutsR/}
}

}
\author{
Sarah Zeller
}
