% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dist.R
\name{extract_dist_n}
\alias{extract_dist_n}
\title{Extract distances and n from donut_analysis model}
\usage{
extract_dist_n(models)
}
\arguments{
\item{models}{List of \code{donut_analysis} models, or single \code{donut_analysis} model}
}
\value{
A tibble
}
\description{
Convenience function to extract distances as well as number of treated,
number of observations, number of original observations, and percentage of
treated from donut_analysis model.
Input needs to be a (list of) \code{donut_analysis} models.
}
\examples{
library(plm)
library(dplyr)
data(Cigar)
Cigar <- Cigar |>  mutate(dist_km = rnorm(nrow(Cigar), 20, 10)) |>  filter(dist_km >= 0)
cigar_model <-
donut_analysis(dist = c(5, 20), ds = Cigar, dep_var = "price", indep_vars = "pop", fe = "state")
extract_dist_n(cigar_model)
}
