% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{extract_info}
\alias{extract_info}
\title{Extract additional info from donut_analysis model}
\usage{
extract_info(models)
}
\arguments{
\item{models}{List of \code{donut_analysis} models, or single \code{donut_analysis} model}
}
\value{
A tibble
}
\description{
Convenience function to extract distances as well as number of treated,
observations, original observations, percentage of treated, clusters, and
treated clusters from donut_analysis model.
Input needs to be a (list of) \code{donut_analysis} models.
}
\examples{
data(donut_data)
model <-
donut_analysis(dist = c(5, 20), ds = donut_data, dep_var = "wealth_index",
indep_vars = "age", fe = "id", clust = TRUE, bootstrap = FALSE)
extract_info(model)
}
