% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info_clust.R
\name{extract_info_clust}
\alias{extract_info_clust}
\title{Helper function for modelsummary}
\usage{
extract_info_clust(donut_model, hist = TRUE, ...)
}
\description{
Extracts infos from \code{donut_list} which are clustered
}
\keyword{internal}
