% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info_conley_basic.R
\name{extract_info_conley_basic}
\alias{extract_info_conley_basic}
\title{Helper function for modelsummary}
\usage{
extract_info_conley_basic(donut_model, ...)
}
\description{
Extracts infos from \code{donut_list} which is not clustered
}
\keyword{internal}
