% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_data.fixest.R
\name{get_p_data.fixest}
\alias{get_p_data.fixest}
\title{Get data on p-values}
\usage{
\method{get_p_data}{fixest}(donut_model, filter_80, name, ...)
}
\arguments{
\item{donut_model}{Object of class \code{donut_model}}

\item{filter_80}{Boolean: should the result be ignored if the treated pop. is >80\%?}

\item{name}{The name of the variable to be plotted}
}
\description{
Helper function, extracts data on p-values for \code{plot_significance}.
}
\keyword{internal}
