% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_data.R
\name{get_p_data}
\alias{get_p_data}
\title{Get data on p-values}
\usage{
get_p_data(donut_model, filter_80, ...)
}
\arguments{
\item{donut_model}{Object of class \code{donut_model}}

\item{filter_80}{Boolean: should the result be ignored if the treated pop. is >80\%?}
}
\description{
Helper function, extracts data on p-values for \code{plot_significance}.
}
\keyword{internal}
