% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_rows.R
\name{info_rows}
\alias{info_rows}
\title{Helper function for modelsummary}
\usage{
info_rows(donut_model, hist = TRUE, footnote = FALSE, ...)
}
\description{
Extracts infos from \code{donut_list}
}
\keyword{internal}
