% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_modelsummary_list.R
\name{make_modelsummary_list}
\alias{make_modelsummary_list}
\title{Helper function for modelsummary}
\usage{
make_modelsummary_list(
  donut_list,
  output,
  hist,
  title,
  vars,
  footnote_numbers,
  footnote_bs,
  ...
)
}
\description{
Makes modelsummary_list with donut_list elements
}
\keyword{internal}
