% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donutsR.R
\docType{package}
\name{myFunctions}
\alias{donutsR}
\alias{donutsR-package}
\title{Sarah's custom functions
Often used functions.}
\description{
With this package, you can run donut analyses: regressions with a distance parameter, which is used to create treatment dummies, and to determine the population.
}
\author{
Sarah Zeller
}
