% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_case.R
\name{sentence_case}
\alias{sentence_case}
\title{Helper function for model_summary}
\usage{
sentence_case(x)
}
\description{
Turns character vectors to sentence case, and retains the names.
}
\keyword{internal}
