% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorted_vars.R
\name{sorted_vars}
\alias{sorted_vars}
\title{Helper function for donut_summary}
\usage{
sorted_vars(donut_list)
}
\description{
Prettifies variables and their labels, and sorts them so "dist" is up front
}
\keyword{internal}
