% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom.fixest.R
\name{tidy_custom.fixest}
\alias{tidy_custom.fixest}
\title{Helper function for modelsummary}
\usage{
\method{tidy_custom}{fixest}(x, ...)
}
\description{
Helper function for modelsummary
}
\keyword{internal}
