% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_analysis.R
\name{donut_analysis}
\alias{donut_analysis}
\title{Perform a Donut FE analysis}
\usage{
donut_analysis(
  dist,
  ds,
  dep_var,
  indep_vars,
  fe = "id",
  dist_var = "dist_km",
  se = "cluster",
  bootstrap = FALSE,
  B = 9999,
  lat = "lat",
  lon = "lon",
  excl_largest_fe = 0,
  excl_inner_r = 0,
  treatment_label = "Lives close to landfill",
  ...
)
}
\arguments{
\item{dist}{A vector with two distances: the inner and the outer.
The inner distance refers to the maximum distance for treatment.
The outer distance refers to the overall population: it is the cut-off value.}

\item{ds}{The dataset.}

\item{dep_var}{The dependent variable in the regression}

\item{indep_vars}{a character vector of independent variables. Must be part of \code{ds}.}

\item{fe}{a character vector of Fixed Effects variables. Must be part of \code{ds}.}

\item{dist_var}{A character referring to the distance parameter.
Defaults to \code{dist_km}.}

\item{se}{is a character, defaulting to \code{cluster}.
It shows how to adjust standard errors: By clustering them by the \code{fe} param
(\code{cluster}), no adjustment, (\code{basic}), or with Conley (\code{conley}).
All regressions rely on \code{fixest}.}

\item{bootstrap}{a boolean: should the standard errors for \code{dist} be bootstrapped?}

\item{B}{the number of bootstraps}

\item{lat}{Latitude; needed for \code{conley}-standard errors}

\item{lon}{Longitude; needed for \code{conley}-standard errors}

\item{excl_largest_fe}{Numeric: how many of the FE with the largest number of
observations should be excluded? Defaults to \code{0}.}

\item{excl_inner_r}{Numeric. If the observations that are closest to the \code{fe}
should be omitted: in which radius should they be omitted? Defaults to \code{0}.}

\item{treatment_label}{Character. The label for the treatment variable created for
the combination of inner and outer radius. Defaults to "Lives close to landfill"}

\item{...}{Additional arguments}
}
\value{
A single \code{donut_model} object with class \code{fixest}.
The \code{donut_model} has additional list element which contain
the inner and outer radii and additional info.
It can be plotted with the \code{plot_significance} function, and showed as table
with \code{donut_summary}.
}
\description{
Performs a single FE-models with a given data set.
The independent variable and the population can be adjusted.
}
\examples{
data(donut_data)
model1 <-
donut_analysis(dist = c(5, 20),
               ds = donut_data,
               dep_var = "wealth_index",
               indep_vars = "age",
               fe = "id")
model2 <-
  donut_analysis(dist = c(8, 50),
                 ds = donut_data,
                 dep_var = "wealth_index",
                 indep_vars = c("age", "male"),
                 fe = "id",
                 dist_var = "dist_km",
                 excl_largest_fe = 2,
                 excl_inner_r = 1)
model_basic <-
  donut_analysis(dist = c(8, 50),
                 ds = donut_data,
                 dep_var = "wealth_index",
                 indep_vars = c("age", "male"),
                 fe = "id",
                 dist_var = "dist_km",
                 se = "basic")
}
