% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_analysis.R
\name{donut_analysis}
\alias{donut_analysis}
\title{Perform a Donut FE analysis}
\usage{
donut_analysis(dist, ds, dep_var, indep_vars, fe = "id")
}
\arguments{
\item{dist}{A vector with two distances: the inner and the outer.
The inner distance refers to the maximum distance for treatment.
The outer distance refers to the overall population: it is the cut-off value.}

\item{ds}{The dataset. A column "dist_km" MUST be part of it.}

\item{dep_var}{The dependent variable in the regression}

\item{indep_vars}{a character vector of independent variables. Must be part of \code{ds}.}

\item{fe}{a character vector of Fixed Effects variables. Must be part of \code{ds}.}
}
\value{
A single plm model with an additional list element which contains
the inner and outer radii.
}
\description{
Performs a single FE-models with a given data set.
The independent variable and the population can be adjusted.
}
\examples{
library(plm)
library(dplyr)
data(Cigar)
Cigar <- Cigar |>  mutate(dist_km = rnorm(nrow(Cigar), 20, 10)) |>  filter(dist_km >= 0)
cigar_model <-
donut_analysis(dist = c(5, 20), ds = Cigar, dep_var = "price", indep_vars = "pop", fe = "state")

}
