% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_analysis.R
\name{donut_analysis}
\alias{donut_analysis}
\title{Perform a Donut FE analysis}
\usage{
donut_analysis(
  dist,
  ds,
  dep_var,
  indep_vars,
  fe = "id",
  dist_var = "dist_km",
  se = "cluster",
  bootstrap = FALSE,
  B = 9999,
  lat = "lat",
  lon = "lon",
  excl_largest_fe = 0,
  ...
)
}
\arguments{
\item{dist}{A vector with two distances: the inner and the outer.
The inner distance refers to the maximum distance for treatment.
The outer distance refers to the overall population: it is the cut-off value.}

\item{ds}{The dataset.}

\item{dep_var}{The dependent variable in the regression}

\item{indep_vars}{a character vector of independent variables. Must be part of \code{ds}.}

\item{fe}{a character vector of Fixed Effects variables. Must be part of \code{ds}.}

\item{dist_var}{A character referring to the distance parameter.
Defaults to \code{dist_km}.}

\item{se}{is a character, defaulting to \code{cluster}.
It shows how to adjust standard errors: By clustering them by the \code{fe} param
(\code{cluster}, fixest::feols),
no adjustment, (\code{plm}, plm::plm), or with \code{conley} (\code{conley} param in fixest::feols).}

\item{bootstrap}{a boolean: should the standard errors for \code{dist} be bootstrapped?}

\item{B}{the number of bootstraps}

\item{lat}{Latitude; needed for \code{conley}-standard errors}

\item{lon}{Longitude; needed for \code{conley}-standard errors}

\item{excl_largest_fe}{Numeric: how many of the FE with the largest number of
observations should be excluded? Defaults to \code{NULL}.}

\item{...}{Additional arguments}
}
\value{
A single \code{donut_model} object with class \code{plm} (clust = FALSE) or
class \code{fixest}. The \code{donut_model} has additional list element which contain
the inner and outer radii and additional info.
It can be plotted with the \code{plot_significance} function.
}
\description{
Performs a single FE-models with a given data set.
The independent variable and the population can be adjusted.
}
\examples{
library(fixest)
library(dplyr)
data(Cigar)
set.seed(123)
Cigar1 <- Cigar |>  mutate(dist_km = stats::rnorm(nrow(Cigar), 20, 10)) |>  filter(dist_km >= 0)
cigar_model1 <-
donut_analysis(dist = c(5, 20), ds = Cigar1, dep_var = "price", indep_vars = "pop", fe = "state")
Cigar2 <- Cigar |>  mutate(dist_km1 = rnorm(nrow(Cigar), 20, 10)) |>  filter(dist_km1 >= 0)
cigar_model2 <-
  donut_analysis(dist = c(5, 20),
                 ds = Cigar2,
                 dep_var = "price",
                 indep_vars = "pop",
                 fe = "state",
                 dist_var = "dist_km1")
}
