% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{extract_info}
\alias{extract_info}
\title{Extract additional info from donut_analysis model}
\usage{
extract_info(models)
}
\arguments{
\item{models}{List of \code{donut_analysis} models, or single \code{donut_analysis} model}
}
\value{
A tibble
}
\description{
Convenience function to extract distances as well as number of treated,
observations, original observations, percentage of treated, clusters, and
treated clusters from donut_analysis model.
Input needs to be a (list of) \code{donut_analysis} models.
}
\examples{
library(plm)
library(dplyr)
data(Cigar)
Cigar <- Cigar |>  mutate(dist_km = rnorm(nrow(Cigar), 20, 10)) |>  filter(dist_km >= 0)
cigar_model <-
donut_analysis(dist = c(5, 20), ds = Cigar, dep_var = "price", indep_vars = "pop", fe = "state")
extract_info(cigar_model)
}
