% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{model_summary}
\alias{model_summary}
\title{Model summary table for donut_list}
\usage{
model_summary(
  donut_list,
  r_inner = NULL,
  filter_80 = TRUE,
  title = NULL,
  hist = TRUE,
  ...
)
}
\arguments{
\item{donut_list}{The donut_list}

\item{r_inner}{The inner radius; optional. A (vector of) integers which are
the inner radius of a \code{donut_model}. Must be in the \code{donut_list}.}

\item{filter_80}{Should only regressions where treated are <=80\% be shown?
Defaults to true.}

\item{title}{Optional: a title for the summary table.
Defaults to "Regression results (r_inner km radius)".}

\item{hist}{If \code{donut_list} has clustered standard errors: Should a histogram of the
clusters be shown?}

\item{...}{Optional arguments for histogram and modelsummary}
}
\value{
A modelsummary object, usually a \code{kableExtra} table.
}
\description{
Convenience function to make a modelsummary table.
Input needs to be a list of \code{donut_analysis} models, i.e. a \code{donut_list}.
}
\examples{
data(donut_data)
models <-
donut_models(inner = 2:4, outer = c(10, 20), ds = donut_data,
dep_var = "wealth_index", indep_vars = "age", fe = "id")
table <-
model_summary(donut_list = models, r_inner = 3, height = 200)

}
