% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\docType{package}
\name{myFunctions}
\alias{myFunctions}
\alias{myFunctions-package}
\title{Sarah's custom functions
Often used functions.}
\description{
These are some functions I often use.
}
\author{
Sarah Zeller
}
