% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_significance.R
\name{plot_significance}
\alias{plot_significance}
\title{Extract distances from donut_analysis model}
\usage{
plot_significance(donut_models, var, filter_80 = TRUE)
}
\arguments{
\item{donut_models}{List of \code{donut_analysis} models}

\item{var}{The dependent variable in all \code{donut_analysis} models which should be plotted.}

\item{filter_80}{Boolean: Should only those regressions be displayed for which
the treatment group is >= 80\% of the whole population? Defaults to \code{TRUE}.}
}
\value{
A ggplot2 plot
}
\description{
Convenience function to extract distances from donut_analysis model.
Input needs to be a (list of) \code{donut_analysis} models.
Consider adding \code{theme(legend.text.align = 1)} to right-align the p-value legend.
}
\examples{
library(plm)
library(dplyr)
data(Cigar)
Cigar <- Cigar |>  mutate(dist_km = rnorm(nrow(Cigar), 20, 10)) |>  filter(dist_km >= 0)
cigar_models <-
donut_models(inner = 2:4, outer = c(10, 20), ds = Cigar,
dep_var = "dist_km", indep_vars = "pop", fe = "state")
plot_significance(cigar_models, var = "pop")
}
